vim-auto-pairs turned out to be bullshit.
In some ways it was an improvement over AutoClose, in that it was
simpler, but that turned out to be no protection against bugginess, and
it interfered with using the return key in VimWiki due to a
poorly-written eval statement. Also, it didn't play nicely with comments
and strings and broke undo. vim-smartinput seems to address those issues
and now vimwiki is working again, so let's see how things go.